Moved files from the modules folder to `src`
